Move A64 files around to make more sense
